More Textures And Optimizations

+++ src/code/World/Chunk/MeshData.ts
@@ -2,10 +2,10 @@ export class MeshData {
   positions: Uint8Array = new Uint8Array();
   indices: Uint16Array = new Uint16Array();
   normals: Int8Array = new Int8Array();
-  tangents: Int8Array = new Int8Array();
-  uvs2: Float32Array = new Float32Array();
-  uvs3: Float32Array = new Float32Array();
-  cornerIds: Uint8Array = new Uint8Array();
+  tangents: Int8Array = new Int8Array(); // vec4
+  uvs2: Uint8Array = new Uint8Array(); // vec2, tile coords (tx, ty)
+  uvs3: Uint8Array = new Uint8Array(); // vec2, quad dimensions (w, h)
+  cornerIds: Uint8Array = new Uint8Array(); // float
 
   public static deserialize(data: any): MeshData {
     const meshData = new MeshData();
@@ -17,8 +17,8 @@ export class MeshData {
     meshData.indices = data.indices || new Uint16Array();
     meshData.normals = data.normals || new Int8Array();
     meshData.tangents = data.tangents || new Int8Array();
-    meshData.uvs2 = data.uvs2 || new Float32Array();
-    meshData.uvs3 = data.uvs3 || new Float32Array();
+    meshData.uvs2 = data.uvs2 || new Uint8Array();
+    meshData.uvs3 = data.uvs3 || new Uint8Array();
     meshData.cornerIds = data.cornerIds || new Uint8Array();
 
     return meshData;
